Faster handling of waiters for the common case of 1 fiber waiting on 1 socket
The catch is that we naively do array appends and all of these
allocate on GC heap and ultimately produce garabage.

Also observe that we shrink slices, D2 mandates that slice doesn't
point to the tail of original array reallocates on append.

The reason is to avoid accidentally stomping on memory used elsewhere.
Safety is critical but it became a typical performance pitfall.